build: rework public symbol exporting logic
With MSVC the symbols are hidden by default, so we need a mechanism
similar to what we do with GNU and Apple linkers to export them. The
simplest way to achieve that is to use some ifdefery to set some
specific import/export attributes.

On Windows, this implies that users will be required to set a special
compilation flag for their static build. To make this transparent for
the users, the specific flag is exposed through pkg-config if the build
is configured for static. pkg-config is arguably not well supported on
Windows, but this is better than nothing. For DLL users, nothing
specific is required since it's the preferred and default build
configuration.

As a side effect, the linker scripts for GNU and Apple are not required
anymore. The meson gnu_symbol_visibility option also works with Apple
Clang/LLVM toolchain.